refactor: put HeaderVecWeak into its own module
Things becoming bigger ;)
